server/requirements.txt
fastapi[all]
psycopg2-binary
python-dotenv
SQLAlchemy==1.4.0
passlib[bcrypt]
python-jose[cryptography]
faker